Suppress sqlite warnings for IPython on initialization
It is more of an ugly hack, required because of django injecting
global callback in a database api.

Ideally, instead of registering adapter for `datetime.datetime`
(django.db.backends.sqlite3.base.py Line 76), django should do it
for a pseudo wrapper over `datetime.datetime` and use it everywhere,
so that other applications and libraries don't have to worry about
the behaviour being infected.

Signed-off-by: Rohan Jain <crodjer@gmail.com>